openfireservice/src/requirements.txt
cloudshell-orch-core>=3.0.0.0,<3.1.0.0
cloudshell-shell-core>=5.0.3,<6.0.0
requests
sleekxmpp